MIDI output config handling
Also changes the way MIDI channels are handled in core. Channels are now
set on the output stream, not the string anymore. And notes don't track
the channel they are on, but simply if they are on or off. core makes
sure that notes are turned off before switching channels.