Rework trompette chien modelling, including new percussive mode
Rework the standard chien modelling by removing a lot of smoothing,
the acceleration calculation and the normalization of the chien volume
range. Makes for a more direct chien, better controllable and with more
"punch". Probably requires adjustment of custom speed to chien volume
mapping.

Also add percussive mode for chien (also easier to work with over MIDI)
and a custom speed to percussion volume mapping.